Move response formatting to the execution service (#1771)
Response formatting was applied at the supergraph service level, which
means that the execution service is producing invalid responses (with
too much data), and that any change to the response in execution
plugins will likely be erased by response formatting.
This moves response formatting to the execution plugin instead.
