+++ apollo-router/src/services/execution_service.rs
@@ -21,7 +21,9 @@ use super::layers::allow_only_http_post_mutations::AllowOnlyHttpPostMutationsLay
 use super::new_service::NewService;
 use super::subgraph_service::SubgraphServiceFactory;
 use super::Plugins;
+use crate::graphql::IncrementalResponse;
 use crate::graphql::Response;
+use crate::json_ext::ValueExt;
 use crate::services::execution;
 use crate::ExecutionRequest;
 use crate::ExecutionResponse;
@@ -59,13 +61,12 @@ where
             let context = req.context;
             let ctx = context.clone();
             let (sender, receiver) = futures::channel::mpsc::channel(10);
-
-            let operation_name = &req.supergraph_request.body().operation_name;
-            let variables = &req.supergraph_request.body().variables;
+            let variables = req.supergraph_request.body().variables.clone();
+            let operation_name = req.supergraph_request.body().operation_name.clone();
 
             let is_deferred = req
                 .query_plan
-                .is_deferred(operation_name.as_deref(), variables);
+                .is_deferred(operation_name.as_deref(), &variables);
 
             let first = req
                 .query_plan
@@ -78,11 +79,73 @@ where
                 )
                 .await;
 
+            let query = req.query_plan.query.clone();
             let stream = if is_deferred {
                 filter_stream(first, receiver).boxed()
             } else {
                 once(ready(first)).chain(receiver).boxed()
             };
+
+            let schema = this.schema.clone();
+
+            let stream = stream
+                .map(move |mut response: Response| {
+                    let has_next = response.has_next.unwrap_or(true);
+                    tracing::debug_span!("format_response").in_scope(|| {
+                        query.format_response(
+                            &mut response,
+                            operation_name.as_deref(),
+                            is_deferred,
+                            variables.clone(),
+                            schema.api_schema(),
+                        )
+                    });
+
+                    match (response.path.as_ref(), response.data.as_ref()) {
+                        (None, _) | (_, None) => {
+                            if is_deferred {
+                                response.has_next = Some(has_next);
+                            }
+
+                            response
+                        }
+                        // if the deferred response specified a path, we must extract the
+                        // values matched by that path and create a separate response for
+                        // each of them.
+                        // While { "data": { "a": { "b": 1 } } } and { "data": { "b": 1 }, "path: ["a"] }
+                        // would merge in the same ways, some clients will generate code
+                        // that checks the specific type of the deferred response at that
+                        // path, instead of starting from the root object, so to support
+                        // this, we extract the value at that path.
+                        // In particular, that means that a deferred fragment in an object
+                        // under an array would generate one response par array element
+                        (Some(response_path), Some(response_data)) => {
+                            let mut sub_responses = Vec::new();
+                            response_data.select_values_and_paths(response_path, |path, value| {
+                                sub_responses.push((path.clone(), value.clone()));
+                            });
+
+                            Response::builder()
+                                .has_next(has_next)
+                                .incremental(
+                                    sub_responses
+                                        .into_iter()
+                                        .map(move |(path, data)| {
+                                            IncrementalResponse::builder()
+                                                .and_label(response.label.clone())
+                                                .data(data)
+                                                .path(path)
+                                                .errors(response.errors.clone())
+                                                .extensions(response.extensions.clone())
+                                                .build()
+                                        })
+                                        .collect(),
+                                )
+                                .build()
+                        }
+                    }
+                })
+                .boxed();
 
             Ok(ExecutionResponse::new_from_response(
                 http::Response::new(stream as _),
